+++ applications/main/infrared/views/infrared_progress_view.c
@@ -62,7 +62,7 @@ static void infrared_progress_view_draw_callback(Canvas* canvas, void* _model) {
         canvas, x + 33, y + 37, AlignCenter, AlignCenter, progress_string);
 
     uint8_t buttons_x = x + (model->is_paused ? 10 : 14);
-    uint8_t buttons_y = y + (model->is_paused ? 46 : 50);
+    uint8_t buttons_y = y + (model->is_paused ? 42 : 50);
 
     canvas_draw_icon(canvas, buttons_x + 0, buttons_y + 0, &I_Pin_back_arrow_10x8);
     canvas_draw_str(canvas, buttons_x + 14, buttons_y + 8, model->is_paused ? "resume" : "stop");
@@ -74,6 +74,9 @@ static void infrared_progress_view_draw_callback(Canvas* canvas, void* _model) {
         canvas_draw_icon(canvas, buttons_x + 2, buttons_y + 21, &I_ButtonLeftSmall_3x5);
         canvas_draw_icon(canvas, buttons_x + 7, buttons_y + 21, &I_ButtonRightSmall_3x5);
         canvas_draw_str(canvas, buttons_x + 14, buttons_y + 26, "select");
+
+        canvas_draw_icon(canvas, buttons_x + 2, buttons_y + 30, &I_SmallArrowDown_3x5);
+        canvas_draw_str(canvas, buttons_x + 14, buttons_y + 35, "save");
     }
 }
 
@@ -128,6 +131,8 @@ bool infrared_progress_view_input_callback(InputEvent* event, void* context) {
                 else if(event->key == InputKeyOk)
                     instance->input_callback(
                         instance->context, InfraredProgressViewInputSendSingle);
+                else if(event->key == InputKeyDown)
+                    instance->input_callback(instance->context, InfraredProgressViewInputSave);
                 else if(event->key == InputKeyBack)
                     instance->input_callback(instance->context, InfraredProgressViewInputResume);
             } else {