+++ applications/main/infrared/views/infrared_progress_view.h
@@ -20,6 +20,7 @@ typedef enum {
     InfraredProgressViewInputPreviousSignal,
     InfraredProgressViewInputNextSignal,
     InfraredProgressViewInputSendSingle,
+    InfraredProgressViewInputSave,
 } InfraredProgressViewInput;
 
 /** Callback for input handling */